LayoutTests/http/tests/resourceLoadStatistics/omit-referrer-for-navigation-ephemeral.html

<!-- webkit-test-runner [ useEphemeralSession=true ] -->
<!DOCTYPE html>
<html>
<head>
    <meta name="referrer" content="unsafe-url">
    <script src="/js-test-resources/js-test.js"></script>
    <script src="/cookies/resources/cookie-utilities.js"></script>
    <script src="resources/util.js"></script>
</head>
<body onload="setTimeout('runTest()', 0)">
<script>
    description("Check that document.referrer is omitted after a cross-site navigation with link query while in private browsing mode.");
    jsTestIsAsync = true;

    let numberOfTries = 0;
    function testReferrer() {
        if (!document.referrer) {
            testPassed("The referrer is omitted.");
            setEnableFeature(false, finishJSTest);
        } else if (++numberOfTries <= 5)
            setTimeout(testReferrer, 200);
        else {
            testFailed("The referrer is not omitted: " + document.referrer);
            setEnableFeature(false, finishJSTest);
        }
    }

    function navigateCrossOrigin() {
        document.location.href = destinationOrigin + "/resourceLoadStatistics/omit-referrer-for-navigation-ephemeral.html";
    }

    const destinationOrigin = "http://localhost:8000";
    const prevalentResourceOrigin = "http://127.0.0.1:8000";
    function runTest() {
        if (document.location.origin === prevalentResourceOrigin) {
            navigateCrossOrigin(); 
        } else {
            testReferrer();
        }
    }
</script>
</body>
</html>
